switched from columns()
switched from columns() (which are fixed width) to the meander package which lets us specify variable-width columns in the template.

Note that there is a feature request pending that will let us go back to columns.  This is preferable because things like colbreak, gt tables, and some other gotchas do not currently work in template content